requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
pandas>=2.0.0
streamlit>=1.28.0
opencv-python>=4.8.0
tensorboard>=2.13.0
kaggle>=1.5.16
timm>=0.9.12

